reorg requests AND components
