+++ src/components/jobcompenents/TargetJobSnapshot.js
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { Link } from 'react-router-dom';
-import { requestDeleteTargetJob } from './Requests'
+import { requestDeleteTargetJob } from '../requests/JobRequests'
 
 
 export const TargetJobSnapshot = ({ token, job }) => {
